+++ var/spack/repos/builtin/packages/foam-extend/package.py
@@ -58,7 +58,7 @@ import shutil
 import os
 
 from spack import *
-from spack.environment import EnvironmentModifications
+from spack.util.environment import EnvironmentModifications
 from spack.pkg.builtin.openfoam_com import OpenfoamArch
 from spack.pkg.builtin.openfoam_com import add_extra_files
 from spack.pkg.builtin.openfoam_com import write_environ
