+++ var/spack/repos/builtin/packages/intel-ipp/package.py
@@ -25,7 +25,7 @@
 import os
 
 from spack import *
-from spack.environment import EnvironmentModifications
+from spack.util.environment import EnvironmentModifications
 
 
 class IntelIpp(IntelPackage):
