+++ var/spack/repos/builtin/packages/intel-parallel-studio/package.py
@@ -26,7 +26,7 @@ import glob
 import os
 
 from spack import *
-from spack.environment import EnvironmentModifications
+from spack.util.environment import EnvironmentModifications
 
 
 class IntelParallelStudio(IntelPackage):
